Bug Fixes
The side navigation was broken as the PHP file for the side navigation was in the wrong subfolder, this was a simple fix and now works as intended.
I have now changed my contact database and my newsfeed database into one netmatters database with the 2 seperate tables in them.
I changed the sql query that retrieved the the news articles to retrieve only the three latest articles tto reduce page load should the database contain a lot of articles. I did this by changing the SELECT query to SELECT * FROM  posts, ORDER BY `date` DESC and LIMIT to 3.
I fixed a layout issue with the news articles on smaller devices.